custom_template
HSFFIG requires custom template for hsc2hs. While it is possible to install it
with the package, it is inconvenient to remember path to it.
Since the template is just an include file, it may be placed
in the very beginning of the generated .hsc file. hsc2hs will be
run with -t /dev/null then.
To hardcode the template in hsffig, a short program, mkhstmpl.hs is compiled
and run prior to package build. It reads the template from the distribution's include directory
and creates a Haskell module Template.hs which contains the only function
which outputs the file text into standard output.

darcs-hash:2dffda511f1676398b5fa7763f4c955d542d529d